stop-loss codons on transcripts without 3' UTRs can be shorter than the original
